+++ packages/app/lib/page/login/import_nsec.dart
@@ -131,8 +131,6 @@ class _ImportNsec extends State<ImportNsec> {
                                       name: name,
                                       prikey: input,
                                       hexPubkey: hexPubkey);
-                              await IdentityService.instance
-                                  .syncProfileFromRelay(newIdentity);
                               EasyLoading.showSuccess('Import successfully');
                               Get.back(result: newIdentity);
                             } catch (e, s) {